PycharmProjects/PythonProjectSecond/SUDNAXI-Trading-Intelligence/requirements.txt
absl-py==2.3.1
albucore==0.0.24
albumentations==2.0.8
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.11
contourpy==1.3.2
curl_cffi==0.12.0
customtkinter==5.2.2
cv2_enumerate_cameras==1.1.15
cycler==0.12.1
Cython==3.1.2
darkdetect==0.8.0
easydict==1.13
filelock==3.18.0
findspark==2.0.1
flatbuffers==25.2.10
fonttools==4.58.5
frozendict==2.4.6
fsspec==2025.5.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.73.1
h11==0.16.0
h5py==3.14.0
httptools==0.6.4
humanfriendly==10.0
idna==3.10
imageio==2.37.0
insightface==0.7.3
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==3.10.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
multitasking==0.0.11
namex==0.1.0
narwhals==1.47.0
networkx==3.5
nltk==3.9.1
numpy==2.1.3
onnx==1.18.0
onnxruntime-gpu==1.22.0
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
optree==0.16.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
peewee==3.18.2
pillow==11.1.0
platformdirs==4.3.8
plotly==6.2.0
prettytable==3.16.0
protobuf==5.29.5
py4j==0.10.9.7
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pygrabber==0.2
pyparsing==3.2.3
pyreadline3==3.5.4
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
PySocks==1.7.1
pyspark==3.5.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rpds-py==0.26.0
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.16.0
selenium==4.33.0
shiboken6==6.9.1
simsimd==6.5.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
streamlit==1.47.0
stringzilla==3.12.5
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
textblob==0.19.0
threadpoolctl==3.6.0
tifffile==2025.6.11
toml==0.10.2
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
yfinance==0.2.65
